[refactor] home uiEvent파일 생성, 클릭이벤트 처리 , uiState 별도 파일 생성
클릭 이벤트를 sealed class 로 묶어 처리하도록 했습니다
