+++ app/src/main/java/com/mashup/zuzu/ui/home/HomeViewModel.kt
@@ -25,32 +25,9 @@ import javax.inject.Inject
 
 /**
  * @Created by 김현국 2022/07/11
- * @Time 3:22 오후
  */
 
-sealed class BestWorldCupUiState {
-    object Loading : BestWorldCupUiState()
-    object Error : BestWorldCupUiState()
-    data class Success(val bestWorldCupList: List<BestWorldCup>) : BestWorldCupUiState()
-}
 
-sealed class RecommendWineUiState {
-    object Loading : RecommendWineUiState()
-    object Error : RecommendWineUiState()
-    data class Success(val recommendWine: Wine) : RecommendWineUiState()
-}
-
-sealed class MainWineUiState {
-    object Loading : MainWineUiState()
-    object Error : MainWineUiState()
-    data class Success(val mainWines: List<Wine>) : MainWineUiState()
-}
-
-sealed class CategoryListUiState {
-    object Loading : CategoryListUiState()
-    object Error : CategoryListUiState()
-    data class Success(val categoryList: List<Category>) : CategoryListUiState()
-}
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
